+++ MiniEngine/Model.cpp
@@ -237,7 +237,7 @@ bool Model::isLineHitModel(const Vector3& start, const Vector3& end, Vector3& mi
 		for (const auto& indexBuffer : mesh.indexBuffer16Array)
 		{
 			//インデックスの数からポリゴンの数を計算する。
-			int numPolygon = indexBuffer.indices.size() / 3;
+			int numPolygon = static_cast<int>(indexBuffer.indices.size()) / 3;
 			for (int polygonNo = 0; polygonNo < numPolygon; polygonNo++)
 			{
 				Vector3 vertexPosition[3];
